Feat: add the verify_user route
Also added automatic verifaction email and message if user is not verified.
Also added password confirmation field in register route and template.
Also changed the styling of the inputs from input-groupt to form-control in both login and register
Also changed the switch from login to register and viceversa to a button in the templates and an event listener in google.js
Also added javascript function to show toasts and errors in script.js.