Feat: add playground
In template created two containers, one for the chatbot settings, the other for message and answer. Plus a modal to add traits to the assistant base prompt.
In javascript the methods to manage the buttons with popover for the active traits, to set a trait as active, to clear the traits and to invoke the chatbot.
In inference.py modified the methods so when in the future we will have chat_histories it will be compatible giving preference to chat histories.
In routes changed the chat route so it interacts with both the chat method from inference and playground.js