Feat: add create_assistant route to save assistants from the playground
In playground routes passing saved assistants and traits, also added env variables for how many slot the admin want the user to have, default values of 5 assistant and 20 traits if not set.
In playground template added useful information about slots available and how to add existing traits, by clicking on the trait saved the button will trigger a tooltip that will tell the user to double click on the button to add the trait. The tooltip will also show the reason behind the value assigned.
In playground.js moved all the elements at the top and the event listeners at the bottom. now added methods and event listeners to add a saved trait to the active assistant and to save the assistant. It won't be possibile for the user to add a trait with the same value. If the user add an existing trait to a new assistant and save the assistant, the trait will not be saved but still added to the new assistant.
In models fixed the Assistant method retrieve_system_prompt.
In base.html and script.js added timeout and effects to toasts.