Feat: add local_app
Due the impossibility, at least on a basic heroku dyno, to install pytorch (with cpu version slug size 3.1Gb), I'm forced to split the app in two and have another server running for text embeddings. The app on heroku will send request to the second server that will have access on the database by itself. The local app is basically a clone with some added features, models.py, celeryconfig.py  and __init__.py will be exact copies.
Added utils.py in custom_assistant with an utility method that will point to the proprietary_hardware route status.
Added utils.py in proprietary_hardware with utility function to initialize the embedding function and get the server status.