Feat: Add collections route
Modified collection route so an user can manage collections and sources from there. Added a bootstrap offcanvas tom manage the sources with nested a bootstrap accordion for the existing sources and a form to upload a new one. In the main section a select with the existing connections, does nothing at the moment, and a form to submit a new collection.
In script.js added wrapper function to fetch data from server. In collections.js added method to add a new source and the event listener that will call the method.
Added new classes in styles.
In routes added also create_collection and create_source routes which will interact with the collection form and the js addNewSource.
In models.py added name and description to the sources table and the __repr__ method
Also changed app upload folder in __init__.py