Feat: add add_source_to_collection route, BackgroundIngestionTask in custom_assistant module and ingest_data route from proprietary_hardware
In add source to collection it creates a background ingestion task with the collection and source ids, then it'll try to post the id of the task to the embedding server (ingest_data route on proprietary hardware server). If the server is down it will pass the task id to the celery enabled task to start a thread in background, in this case a while loop until the server is up to pass the embedding task.
When finally the ingest_data route on proprietary hardware is called it will start the ingestion process in the background, update the background ingestion task with the proprietary hardware task it.
In the background the server will download the file to ingest from the aws bucket, and extract the pdf into elements, including images and tables which will be uploaded as well to the aws bucket in order to (maybe) use them later in the retrieving and generating process. It will then download the vectorstore from the aws bucket and ingest them in batches (too many will fail). Once ingested it will update the task and the user in the database and upload the vectorstore and all the images extracted to the aws bucket ready to be retrieved at need.
Added html and js to collections to start the process.