Feat: add query_retriever to chat route
In collections.html added the chat-container from playground.html with a nested card for the answer.
In collections.js added the chat and showAnswer method from playground.js and modified accordingly.
In chat custom_assistant route added code to accept the previously mentioned js chat method and act accordingly.
In custom_assistant inference chat added code to post a request to the proprietary hardware route query with question and collection id.
In the proprietary hardware route call the vectorstore query_with_retriever method which will download the client for the given user (it will stay persinstent until the heroku dyno go to sleep due inactivity), it will initialize the vectorstore as a retriever and create a chain which will return the answer to question about the context from the llm as a string, more about the chain in readme.
The answer will be forwarded back as a response from the proprietary_hardware query route.
In proprietary_hardware vectorstore also removed code i wanted to use for the completion when the server is off and build the prompt as i would for an openai compatible api, not realizing that i was already in the proprietary server. I will disable the collection query when the embedding server is off, while it still be possible to ingest sources into collections, or better said, ingest the documents later when the server goes back up thanks to the celery background workers.
Modified playground html and js to style and act accordingly like collections.