Feat: add assistants/collections management system
Used collection.html as template for assistants.html. In the assistants offcanvas the user can manage the saved traits and add new ones.
Added modals in both collection.html and assistants.html to confirm record deletion.
Added routes to delete collections, sources, assistants and traits.
Added routes get_assistants and get_assistant.
Select elements in both assistants.html and collections.html interact with the get_assistant and get_collection routes which will return the data associated with the selected records. If traits or sources are present in the selected assistant or collection, it will create a button with a popover for every record found. It also make the inputs and the save button disabled, it will hide the clear button and make visible the delete and edit buttons. Clicking the edit button will make the delete button disapper and the clear button come back. If the clear button is clicked it sets back the inputs from the active assistant/collection that was set by the select event in the first place. It will also disable the inputs, hide the clear button once again and come back the delete. Inthe offcanvas the user can add the source to an active collection. The user should not be supposed the delete a source without deleting the collection, will fix it later, more on readme.
In vectorstore fixed missing file_name declaration after an OperationalError.
In utils added method to send the user an email when the ingestion process is done.