Feat: add token usage collections and sources to profile page
In chat route now it creates or updates the daily tokens usage for the current user.
In delete source route now it does not let the user delete a source until there are collections with it ingested. More in readme.
In profile route now displays the daily usage with a selector for the day via api. also displays the sources dividing them in deletable or not. I didn't change anything in collections.html template, the delete_source route will send a proper flash response. The profile template now displays tooltips on collections and sources. Moved the js updateTooltips method from playground.js to script.js. Added modal to delete the account with a typing confirmation request.
In base template connected profile page to profile button