Maint: fix the second and following bot answers
It seems that the ollama server running on localhost:11434/v1/chat/completions, as an openai compatible url, when i invoke it passing the chat history it starts with <|im_start|>assistant<|im_end|>, very annoying.. plus it loses the attributes given very fast.
I created a route in proprietary_hardware to which the heroku app will post requests, it will redirect to the history_chat utility method. It instantiate the ChatOllama class, and invoke it with the chat history after converting it.